Fix most of 64 bit to 32 bit pointer truncations There are some harder ones left

+++ src/dynamic_placeholder_text_util.c
@@ -17,7 +17,7 @@ void DynamicPlaceholderTextUtil_Reset(void)
     do
     {
         *ptr2-- = fillval;
-    } while ((int)ptr2 >= (int)ptr);
+    } while (ptr2 >= ptr);
 }
 
 void DynamicPlaceholderTextUtil_SetPlaceholderPtr(u8 idx, const u8 *ptr)
