+++ src/pokemon_icon.c
@@ -1191,7 +1191,7 @@ const u8 *GetMonIconTiles(u16 species, bool32 handleDeoxys)
     const u8 *iconSprite = gMonIconTable[species];
     if (species == SPECIES_DEOXYS && handleDeoxys == TRUE)
     {
-        iconSprite = (const u8 *)(0x400 + (u32)iconSprite); // use the specific Deoxys form icon (Speed in this case)
+        iconSprite = (const u8 *)(0x400 + iconSprite); // use the specific Deoxys form icon (Speed in this case)
     }
     return iconSprite;
 }
